[buteo-sync-plugins-social] Implement rudimentary connection-loss handling. Contributes to MER#885
This commit adds basic handling for connection loss during sync.
In most cases, it detects the connection loss and then skips any
database commit during finalisation.

In some cases where this is not possible, it commits as much data
as possible to ensure that the device remains usable, and then marks
the account for clean sync on next cycle.

Contributes to MER#885

+++ src/facebook/facebook-notifications/facebooknotificationsyncadaptor.cpp
@@ -58,9 +58,13 @@ void FacebookNotificationSyncAdaptor::beginSync(int accountId, const QString &ac
 void FacebookNotificationSyncAdaptor::finalize(int accountId)
 {
     Q_UNUSED(accountId)
-    m_db.purgeOldNotifications(OLD_NOTIFICATION_LIMIT_IN_DAYS);
-    m_db.sync();
-    m_db.wait();
+    if (syncAborted()) {
+        SOCIALD_LOG_INFO("sync aborted, won't commit database changes");
+    } else {
+        m_db.purgeOldNotifications(OLD_NOTIFICATION_LIMIT_IN_DAYS);
+        m_db.sync();
+        m_db.wait();
+    }
 }
 
 void FacebookNotificationSyncAdaptor::requestNotifications(int accountId, const QString &accessToken, const QString &until, const QString &pagingToken)
